windows backup

+++ ecstasy/src/geometry/box.cpp
@@ -10,10 +10,10 @@
 #include <filament/RenderableManager.h>
 #include <filament/TransformManager.h>
 
-ecstasy::box::box(filament::Engine& _filament_engine, Eigen::Vector3d _dimention,
+ecstasy::box::box(filament::Engine& _filament_engine, Eigen::Vector3f _dimention,
                   filament::Material const* _material, Eigen::Vector3d _linear_color, bool _culling)
     : filament_engine_(_filament_engine), material_(_material) {
-    const Eigen::Vector3d half_dim = _dimention / 2.0;
+    const Eigen::Vector3f half_dim = _dimention / 2.0;
     auto& hd = half_dim;
 
     vertices_ = {
